# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "BookBlend API",
    "description": "API Docs for bookblend.io",
    "version": "1.0.0"
  },
  "paths": {
    "/books": {
      "get": {
        "summary": "Get Books",
        "description": "Returns a list of books for a specific Goodreads user and shelf.",
        "operationId": "get_books_books_get",
        "parameters": [
          {
            "name": "user_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "The Goodreads user ID.",
              "title": "User Id"
            },
            "description": "The Goodreads user ID."
          },
          {
            "name": "shelf",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "The Goodreads shelf name (e.g., 'read', 'to-read').",
              "default": "all",
              "title": "Shelf"
            },
            "description": "The Goodreads shelf name (e.g., 'read', 'to-read')."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/user": {
      "get": {
        "summary": "Get User Info",
        "description": "Fetches metadata about a Goodreads user, including name and profile information.",
        "operationId": "get_user_info_user_get",
        "parameters": [
          {
            "name": "user_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "The Goodreads user ID.",
              "title": "User Id"
            },
            "description": "The Goodreads user ID."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/blend": {
      "get": {
        "summary": "Get Blend",
        "description": "Generates shared book insights between two Goodreads users.",
        "operationId": "get_blend_blend_get",
        "parameters": [
          {
            "name": "user_id1",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "The Goodreads user ID for the first user.",
              "title": "User Id1"
            },
            "description": "The Goodreads user ID for the first user."
          },
          {
            "name": "user_id2",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "The Goodreads user ID for the second user.",
              "title": "User Id2"
            },
            "description": "The Goodreads user ID for the second user."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  },
  "servers": [
    {
      "url": "https://book-blend-backend.vercel.app/"
    }
  ]
}